pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccs-monitoring"
version = "0.1.0"
description = "Synthetic-to-field benchmark and hybrid baselines for reliable 4D CCS monitoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "scikit-learn>=1.5",
  "scipy>=1.13",
  "torch>=2.0",
]

[project.optional-dependencies]
jax = [
  "jax>=0.9",
  "jaxlib>=0.9",
]
viz = [
  "imageio>=2.37",
  "plotly>=6.0",
  "zarr>=3.1",
]

[project.scripts]
ccs-monitor = "ccs_monitoring.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
